pyproject.toml
[project]
name = "chess-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.2.0",
    "pygame>=2.6.1",
    "python-chess>=1.999",
    "python-dotenv>=1.1.1",
]
